Move show_versions function to _show_versions.py
Extracted _get_clib_info, _get_module_version, _get_ghostscript_version, _check_ghostscript_version functions out of show_versions(). Moved some Python standard library imports to the top-level instead of nesting in the function.
